--- templates/balance_view.html
@@ -1,8 +1,0 @@
-{% extends "base.html" %}
-{% block content %}
-<div class="container mt-5">
-    <h3>Jūsų balansas: {{ balance }} €</h3>
-    <a href="{{ url_for('user.add_balance') }}" class="btn btn-primary mt-3">Papildyti balansą</a>
-</div>
-{% endblock %}
-
