+++ templates/user/balance_topup.html
@@ -1,0 +1,49 @@
+{% extends "base.html" %}
+
+{% block title %}Balanso papildymas{% endblock %}
+
+{% block content %}
+<div class="flex items-center justify-center py-12 px-4 sm:px-6 lg:px-8">
+    <div class="w-full max-w-md p-8 space-y-6 bg-white rounded-lg shadow-md">
+        <div class="text-center">
+            <h2 class="text-3xl font-extrabold text-gray-900">
+                Balanso papildymas
+            </h2>
+            <p class="mt-2 text-sm text-gray-600">
+                Jūsų dabartinis balansas: 
+                <span class="font-bold text-green-600">{{ "%.2f"|format(current_user.user_balance or 0) }} €</span>
+            </p>
+        </div>
+
+        <form class="mt-8 space-y-6" method="POST" action="">
+            {{ form.hidden_tag() }}
+            
+            <div>
+                {{ form.amount.label(class="block text-sm font-medium text-gray-700") }}
+                <div class="mt-1 relative rounded-md shadow-sm">
+                    <div class="absolute inset-y-0 left-0 pl-3 flex items-center pointer-events-none">
+                        <span class="text-gray-500 sm:text-sm"> € </span>
+                    </div>
+                    {{ form.amount(class="block w-full pl-7 pr-12 sm:text-sm border-gray-300 rounded-md focus:ring-green-500 focus:border-green-500", placeholder="0.00") }}
+                </div>
+                <small class="text-gray-500 text-xs mt-1">* Minimalus papildymas yra 0.01 €</small>
+                {% for error in form.amount.errors %}
+                    <span class="text-red-500 text-xs">{{ error }}</span>
+                {% endfor %}
+            </div>
+
+            <div>
+                <button type="submit" class="w-full flex justify-center py-2 px-4 border border-transparent rounded-md shadow-sm text-sm font-medium text-white bg-green-600 hover:bg-green-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-green-500 cursor-pointer">
+                    Papildyti
+                </button>
+            </div>
+        </form>
+        
+        <div class="mt-4">
+             <a href="{{ url_for('user.dashboard') }}" class="w-full flex justify-center py-2 px-4 border border-gray-300 rounded-md shadow-sm text-sm font-medium text-gray-700 bg-white hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500">
+                Grįžti atgal
+            </a>
+        </div>
+    </div>
+</div>
+{% endblock %}
