+++ templates/user/cart.html
@@ -1,0 +1,11 @@
+{% extends "base.html" %}
+
+{% block content %}
+<div class="bg-white p-8 rounded-lg shadow-lg text-center">
+    <h1 class="text-4xl font-bold text-gray-800 mb-4">Čia bus jūsų pirkinių krepšelis</h1>
+    <p class="text-gray-600 mb-6">Geriausios prekės už geriausią kainą greitai bus jūsų krepšelyje.</p>
+    <a href="#" class="bg-green-500 text-white font-bold py-2 px-4 rounded-lg hover:bg-green-600 transition duration-300">  
+        Žiūrėti prekes
+    </a>
+</div>
+{% endblock %}
