--- templates/user_dashboard.html
@@ -1,44 +1,0 @@
-{% extends "base.html" %}
-
-{% block title %}Vartotojo puslapis{% endblock %}
-
-{% block content %}
-  <div style="padding-top: 6rem; max-width: 800px; margin: 0 auto;">
-        <!-- Pranešimų blokas -->
-    {% with messages = get_flashed_messages(with_categories=true) %}
-      {% if messages %}
-        {% for category, message in messages %}
-        <div
-        class="{% if category == 'success' %}alert-success
-           {% elif category == 'error' %}alert-error
-           {% else %}alert-warning{% endif %}"
-        style="padding: 10px; border-radius: 5px; margin-bottom: 10px;"
-        >
-        {{ message }}
-            </div>
-          {% endfor %}
-        </div>
-      {% endif %}
-    {% endwith %}
-    <h1>Sveiki, {{ user.user_name }} {{ user.user_lastname }}!</h1>
-
-    <!-- Balansas ir papildymo mygtukas kampe -->
-    <div style="position: fixed; top: 80px; right: 10px; background: transparent; padding: 12px; border: 1px solid #ccc; border-radius: 6px; box-shadow: 0 0 8px rgba(0,0,0,0.1); text-align: center; z-index: 1100;">
-      <p><strong>Jūsų balansas:</strong> {{ "%.2f"|format(user.user_balance or 0) }} €</p>
-      <a href="{{ url_for('user.add_balance') }}">
-        <button type="button" style="cursor:pointer; padding: 6px 12px; border-radius: 4px; border: none; background-color: #dc3545; color: white;">
-          Papildyti balansą
-        </button>
-      </a>
-    </div>
-
-    <!-- Atsijungimo mygtukas su margin-top -->
-    <div style="margin-top: 3rem;">
-      <a href="{{ url_for('logout.logout') }}">
-        <button type="button" style="cursor:pointer; padding: 6px 12px; border-radius: 4px; border: 1px solid #dc3545; background-color: #dc3545; color: white;">
-          Atsijungti
-        </button>
-      </a>
-    </div>
-  </div>
-{% endblock %}
